+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 **0.7.4**
 
 * AngularClient - Create project - User field rating and user element cell default values fix
-* Minor
+* DataObjects - Element - SortOrder field added
+* Minor updates
 
 **0.7.3**
 
@@ -12,7 +13,7 @@
 **0.7.2**
 
 * WebAPI - In "Debug" mode, AllowAnyOrigin is true
-* Minor
+* Minor updates
 
 **0.7.1**